Byte från int till long och sen la till unitconversion

+++ common/src/main/java/org/example/Transfer.java
@@ -10,7 +10,7 @@ import java.time.LocalDate;
 public class Transfer {
     private String fromAccountNumber;
     private String toAccountNumber;
-    private int amount;
+    private long amount;
     private LocalDate transferDate;
     private TransferStatus status;
 
@@ -30,7 +30,7 @@ public class Transfer {
      * @param amount            ören som ska föras över
      * @param transferDate      vilket datum överföringen ska görasgi
      */
-    public Transfer(String fromAccountNumber, String toAccountNumber, int amount, LocalDate transferDate) {
+    public Transfer(String fromAccountNumber, String toAccountNumber, long amount, LocalDate transferDate) {
         this.fromAccountNumber = fromAccountNumber;
         this.toAccountNumber = toAccountNumber;
         this.amount = amount;
@@ -46,7 +46,7 @@ public class Transfer {
         return toAccountNumber;
     }
 
-    public int getAmount() {
+    public long getAmount() {
         return amount;
     }
